Move tests to separate folder
